// src/crypto/crypto.enum.ts
export enum Crypto {
  HASH_ALGORITHM = 'sha512',
  ENCRYPT_INPUT_ENCODING = 'utf-8',
  ENCRYPT_OUTPUT_ENCODING = 'base64',
  DECRYPT_INPUT_ENCODING = ENCRYPT_OUTPUT_ENCODING,
  DECRYPT_OUTPUT_ENCODING = ENCRYPT_INPUT_ENCODING,
  HASH_INPUT_ENCODING = ENCRYPT_INPUT_ENCODING,
  HASH_DIGEST_ENCODING = 'hex'
}